+++ socketio/src/asynchronous/client/mod.rs
@@ -1,5 +1,5 @@
 mod ack;
+pub(crate) mod async_client;
 pub(crate) mod builder;
 #[cfg(feature = "async-callbacks")]
 mod callback;
-pub(crate) mod client;
